Ignore embedded repo C-DSA

+++ 08_array.cpp
@@ -2,7 +2,7 @@
 using namespace std;
 int main()
 {
-    int arr[5];
+    double arr[5];
     for (int i = 0; i < 5; i++)
     {
         cin >> arr[i];
